modify handler with location
