+++ src/modules/cities/city_repository.ts
@@ -17,17 +17,11 @@ export namespace City {
     ))`
   }
 
-  const getWherePolygon = (requestQuery: Entity.RequestQuery) => `ST_CONTAINS(ST_GEOMFROMTEXT('${getPolygon(requestQuery)}'), location)`
+  const getWherePolygon = (requestQuery: Entity.RequestQuery) => {
+    const wherePolygon = `ST_CONTAINS(ST_GEOMFROMTEXT('${getPolygon(requestQuery)}'), location)`
 
-  const pointRegexRule = (point: string) => {
-    const pointRegex = /^(-?\d+(\.\d+)?),\s*(-?\d+(\.\d+)?)$/
-    return pointRegex.test(point)
+    return wherePolygon
   }
-
-  const isRequestBounds = (requestQuery: Entity.RequestQuery) => requestQuery?.bounds?.ne
-    && requestQuery?.bounds?.sw
-    && pointRegexRule(requestQuery.bounds.ne)
-    && pointRegexRule(requestQuery.bounds.sw)
 
   export const Cities = () => database<Entity.Struct>('cities')
 
@@ -41,8 +35,7 @@ export namespace City {
       .where('is_active', true)
       .orderBy('name', 'asc')
 
-    if (requestQuery.name) query.where('name', 'LIKE', `%${requestQuery.name}%`)
-    if (isRequestBounds(requestQuery)) query.whereRaw(getWherePolygon(requestQuery))
+    query.whereRaw(getWherePolygon(requestQuery))
 
     return query
   }
